Install fee schedule for 01-Apr-2025
At present, the fee schedule supports only a single day, and that day
must be in the future relative to the actual date on which any testing
or demonstration takes place (ref.
https://support.onefact.net/issues/11980).  While both CarParkTestbench
and InteractiveTesting do not know or care which day is supported by the
fee schedule, any interactive testing or demonstration of the system
using browser-based clients must use the TestControl client to set the
date to the same day as that supported by the fee schedule.  This commit
makes that date 01-Apr-2025.