Adjusted date of stay for interactive test bench
The interactive test bench (which is now poorly named because it is
mostly automated)specifies an absolute date and time for the beginning
of the stay, and this must fall within the dates/times supported by the
fee schedule specified for the application.  At present, the application
uses a fee schedule supporting a single day, 01-Apr-2025, so this commit
makes the date of the stay specified by the interactive test bench the
same.